feat : 중첩 그래프 적용 및 windowInset 올바르게 적용

+++ feature/postdetail/src/main/java/com/withpeace/withpeace/feature/postdetail/navigation/PostDetailNavigation.kt
@@ -14,7 +14,7 @@ const val POST_DETAIL_ROUTE_WITH_ARGUMENT =
     "$POST_DETAIL_ROUTE/{$POST_DETAIL_ID_ARGUMENT}/"
 
 fun NavController.navigateToPostDetail(
-    postId: Int,
+    postId: Long,
     navOptions: NavOptions? = null,
 ) = navigate(route = "$POST_DETAIL_ROUTE/$postId", navOptions)
